+++ Changes.md
@@ -4,6 +4,7 @@ Version Information
 Version 400.0.2
 ---------------
   1. Add 'showhome' setting.
+  2. Add 'hvpcustomcss' setting.
 
 Version 400.0.1
 ---------------